+++ src/js/Chat.js
@@ -1,4 +1,4 @@
-import { html, render, useEffect, useState, Component } from './lib/htm.preact.js';
+import { html, Component } from './lib/htm.preact.js';
 import {publicState, localState, showMenu, resetView, activeChat, activeProfile} from './Main.js';
 import { translate as t } from './Translation.js';
 import Helpers from './Helpers.js';
@@ -11,176 +11,7 @@ import QRScanner from './QRScanner.js';
 import VideoCall from './VideoCall.js';
 
 const chats = window.chats = {};
-const autolinker = new Autolinker({ stripPrefix: false, stripTrailingSlash: false});
 const notificationServiceUrl = 'https://iris-notifications.herokuapp.com/notify';
-
-const submitButton = html`
-  <button type="submit">
-    <svg class="svg-inline--fa fa-w-16" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px" viewBox="0 0 486.736 486.736" style="enable-background:new 0 0 486.736 486.736;" xml:space="preserve" width="100px" height="100px" fill="#000000" stroke="#000000" stroke-width="0"><path fill="currentColor" d="M481.883,61.238l-474.3,171.4c-8.8,3.2-10.3,15-2.6,20.2l70.9,48.4l321.8-169.7l-272.4,203.4v82.4c0,5.6,6.3,9,11,5.9 l60-39.8l59.1,40.3c5.4,3.7,12.8,2.1,16.3-3.5l214.5-353.7C487.983,63.638,485.083,60.038,481.883,61.238z"></path></svg>
-  </button>`;
-
-const attachFileButton = html`
-  <button type="button" id="attach-file">
-    <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path fill="currentColor" d="M21.586 10.461l-10.05 10.075c-1.95 1.949-5.122 1.949-7.071 0s-1.95-5.122 0-7.072l10.628-10.585c1.17-1.17 3.073-1.17 4.243 0 1.169 1.17 1.17 3.072 0 4.242l-8.507 8.464c-.39.39-1.024.39-1.414 0s-.39-1.024 0-1.414l7.093-7.05-1.415-1.414-7.093 7.049c-1.172 1.172-1.171 3.073 0 4.244s3.071 1.171 4.242 0l8.507-8.464c.977-.977 1.464-2.256 1.464-3.536 0-2.769-2.246-4.999-5-4.999-1.28 0-2.559.488-3.536 1.465l-10.627 10.583c-1.366 1.368-2.05 3.159-2.05 4.951 0 3.863 3.13 7 7 7 1.792 0 3.583-.684 4.95-2.05l10.05-10.075-1.414-1.414z"/></svg>
-  </button>`;
-
-const emojiPickerButton = html`
-  <button type="button" id="emoji-picker">
-    <svg aria-hidden="true" focusable="false" data-prefix="far" data-icon="smile" class="svg-inline--fa fa-smile fa-w-16" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 496 512"><path fill="currentColor" d="M248 8C111 8 0 119 0 256s111 248 248 248 248-111 248-248S385 8 248 8zm0 448c-110.3 0-200-89.7-200-200S137.7 56 248 56s200 89.7 200 200-89.7 200-200 200zm-80-216c17.7 0 32-14.3 32-32s-14.3-32-32-32-32 14.3-32 32 14.3 32 32 32zm160 0c17.7 0 32-14.3 32-32s-14.3-32-32-32-32 14.3-32 32 14.3 32 32 32zm4 72.6c-20.8 25-51.5 39.4-84 39.4s-63.2-14.3-84-39.4c-8.5-10.2-23.7-11.5-33.8-3.1-10.2 8.5-11.5 23.6-3.1 33.8 30 36 74.1 56.6 120.9 56.6s90.9-20.6 120.9-56.6c8.5-10.2 7.1-25.3-3.1-33.8-10.1-8.4-25.3-7.1-33.8 3.1z"></path></svg>
-  </button>`;
-
-const subscribedToMsgs = {};
-
-class ChatView extends Component {
-  constructor() {
-    super();
-    this.state = {msgListContent: []};
-  }
-
-  componentDidMount() {
-    localState.get('activeChat').on(activeChatId => {
-      this.setState({});
-      if (activeChatId && !subscribedToMsgs[activeChatId]) {
-        const iv = setInterval(() => {
-          if (chats[activeChatId]) {
-            clearInterval(iv);
-            this.subscribeToMsgs(activeChatId);
-            this.setState({});
-          }
-        }, 1000);
-        subscribedToMsgs[activeChatId] = true;
-      }
-    });
-  }
-
-  subscribeToMsgs(pub) {
-    subscribedToMsgs[pub] = true;
-    chats[pub].getMessages((msg, info) => {
-      if (chats[pub].messageIds[msg.time + info.from]) return;
-      msg.info = info;
-      msg.selfAuthored = info.selfAuthored;
-      msg.timeStr = msg.time;
-      msg.time = new Date(msg.time);
-      chats[pub].messageIds[msg.time + info.from] = true;
-      chats[pub].sortedMessages.push(msg);
-      chats[pub].sortedMessages = chats[pub].sortedMessages.sort((a, b) => a.time - b.time);
-      if (!info.selfAuthored && msg.time > (chats[pub].myLastSeenTime || -Infinity)) {
-        if (activeChat !== pub || document.visibilityState !== 'visible') {
-          Notifications.changeChatUnseenCount(pub, 1);
-        }
-      }
-      if (!info.selfAuthored && msg.time > chats[pub].theirLastSeenTime) {
-        chats[pub].theirLastSeenTime = msg.time;
-        lastSeenTimeChanged(pub);
-      }
-      if (!chats[pub].localLatest || msg.timeStr > chats[pub].localLatest.timeStr) {
-        localState.get('chats').get(pub).get('latestTime').put(msg.timeStr);
-        localState.get('chats').get(pub).get('latest').get('time').put(msg.timeStr);
-        localState.get('chats').get(pub).get('latest').get('text').put(msg.text);
-        chats[pub].localLatest = msg;
-        var text = msg.text || '';
-        if (msg.attachments) {
-          text = '['+ t('attachment') +']' + (text.length ? ': ' + text : '');
-        } else {
-          text = msg.text;
-        }
-        if (chats[pub].uuid && !msg.selfAuthored && msg.info.from && chats[pub].participantProfiles[msg.info.from].name) {
-          text = chats[pub].participantProfiles[msg.info.from].name + ': ' + text;
-        }
-        if (info.selfAuthored) {
-          //latestEl.prepend($(seenIndicatorHtml));
-          setLatestSeen(pub);
-          setLatestCheckmark(pub);
-        }
-      }
-      Notifications.notifyMsg(msg, info, pub);
-      if (activeChat === pub) {
-        this.setState({});
-      }
-    });
-  }
-
-  componentDidUpdate() {
-    Helpers.scrollToMessageListBottom();
-  }
-
-  render() {
-    if (!activeChat || !chats[activeChat].sortedMessages) {
-      return html``;
-    }
-
-    const now = new Date();
-    const nowStr = now.toLocaleDateString();
-    let previousDateStr;
-    let previousFrom;
-    const msgListContent = [];
-    if (chats[activeChat].sortedMessages) {
-      Object.values(chats[activeChat].sortedMessages).forEach(msg => {
-        const date = typeof msg.time === 'string' ? new Date(msg.time) : msg.time;
-        if (date) {
-          const dateStr = date.toLocaleDateString();
-          if (dateStr !== previousDateStr) {
-            var separatorText = iris.util.getDaySeparatorText(date, dateStr, now, nowStr);
-            msgListContent.push(html`<div class="day-separator">${t(separatorText)}</div>`);
-          }
-          previousDateStr = dateStr;
-        }
-
-        const from = msg.info.from;
-        let showName = false;
-        if (previousFrom && (from !== previousFrom)) {
-          msgListContent.push(html`<div class="from-separator"/>`);
-          showName = true;
-        }
-        previousFrom = from;
-        msgListContent.push(html`<${Message} ...${msg} showName=${showName} chatId=${activeChat}/>`);
-      });
-    }
-
-    return html`
-      <div class="main-view" id="message-view">
-        <div id="message-list">${msgListContent}</div>
-        <div id="attachment-preview" class="hidden"></div>
-      </div>
-      <div id="not-seen-by-them" style="display: none">
-        <p dangerouslySetInnerHTML=${{ __html: t('if_other_person_doesnt_see_message') }}></p>
-        <p><button class="copy-chat-link">${t('copy_your_chat_link')}</button></p>
-      </div>
-      <div class="message-form" style="display:none">
-        <form autocomplete="off">
-          <input name="attachment-input" type="file" class="hidden" id="attachment-input" accept="image/*" multiple/>
-          ${attachFileButton}
-          ${emojiPickerButton}
-          <input id="new-msg" type="text" placeholder="${t('type_a_message')}" autocomplete="off" autocorrect="off" autocapitalize="sentences" spellcheck="off"/>
-          ${submitButton}
-        </form>
-      </div>`;
-    }
-}
-
-const NewChat = () => html`
-  <div class="main-view" id="new-chat">
-    <h3>${t('have_someones_chat_link')}</h3>
-    <input id="paste-chat-link" type="text" placeholder="${t('paste_their_chat_link')}"/>
-    <button id="scan-chatlink-qr-btn">${t('or_scan_qr_code')}</button>
-    <video id="chatlink-qr-video" width="320" height="320" style="object-fit: cover;"></video>
-    <h3>${t('give_your_chat_link')}</h3>
-    <button class="copy-chat-link">${t('copy_your_chat_link')}</button>
-    <button id="show-my-qr-btn">${t('or_show_qr_code')}</button>
-    <p id="my-qr-code" class="qr-container" style="display:none"></p>
-    <p><small dangerouslySetInnerHTML=${{ __html: t('beware_of_sharing_chat_link_publicly') }}></small></p>
-    <h3>${t('new_group')}</h3>
-    <p>
-      <form id="new-group-form">
-        <input id="new-group-name" type="text" placeholder="${t('group_name')}"/>
-        <button type="submit">${t('create')}</button>
-      </form>
-    </p>
-    <hr/>
-    <h3>${t('your_chat_links')}</h3>
-    <p><button id="generate-chat-link">${t('create_new_chat_link')}</button></p>
-    <div id="my-chat-links" class="flex-table"></div>
-  </div>`;
 
 function showChat(pub) {
   if (!pub) {
@@ -282,51 +113,6 @@ async function onMsgFormSubmit(event, pub) {
   $('#new-msg').val('');
 }
 
-const seenIndicator = html`<span class="seen-indicator"><svg version="1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 59 42"><polygon fill="currentColor" points="40.6,12.1 17,35.7 7.4,26.1 4.6,29 17,41.3 43.4,14.9"></polygon><polygon class="iris-delivered-checkmark" fill="currentColor" points="55.6,12.1 32,35.7 29.4,33.1 26.6,36 32,41.3 58.4,14.9"></polygon></svg></span>`;
-
-class Message extends Component {
-  constructor() {
-    super();
-    this.i = 0;
-  }
-
-  render() {
-    if (++this.i > 1) console.log(this.i);
-    let name, color;
-    // TODO: public messages
-    if (chats[this.props.chatId].uuid && !this.props.info.selfAuthored) {
-      const profile = chats[this.props.chatId].participantProfiles[this.props.info.from];
-      name = profile && profile.name;
-      color = profile && profile.color;
-      if (name) {
-        var nameEl = $('<small>').click(() => addMention(name)).text(name).css({color}).addClass('msgSenderName');
-      }
-    }
-    const emojiOnly = this.props.text.length === 2 && Helpers.isEmoji(this.props.text);
-
-    const p = document.createElement('p');
-    p.innerText = this.props.text;
-    const h = emojiOnly ? p.innerHTML : Helpers.highlightEmoji(p.innerHTML);
-    const innerHTML = autolinker.link(h);
-
-    return html`
-      <div class="msg ${this.props.selfAuthored ? 'our' : 'their'}">
-        <div class="msg-content">
-          ${name && this.props.showName && html`<small onclick=${() => addMention(name)} class="msgSenderName" style="color: ${color}">${name}</small>`}
-          ${this.props.attachments && this.props.attachments.map(a =>
-            html`<img src=${a.data} onclick=${e => { Gallery.openAttachmentsGallery(this.props, e); }}/>` // escape a.data
-          )}
-          <div class="text ${emojiOnly && 'emoji-only'}" dangerouslySetInnerHTML=${{ __html: innerHTML }}>
-          </div>
-          <div class="time">
-            ${iris.util.formatTime(this.props.time)}
-            ${this.props.selfAuthored && seenIndicator}
-          </div>
-        </div>
-      </div>`;
-  }
-}
-
 function deleteChat(pub) {
   iris.Channel.deleteChannel(publicState, Session.getKey(), pub);
   if (activeChat === pub) {
@@ -335,11 +121,6 @@ function deleteChat(pub) {
   }
   delete chats[pub];
   $('.chat-item[data-pub="' + pub +'"]').remove();
-}
-
-function addMention(name) {
-  $('#new-msg').val($('#new-msg').val().trim() + ` @${name} `);
-  $('#new-msg').focus();
 }
 
 function newChat(pub, chatLink) {
@@ -562,6 +343,7 @@ function showNewChat() {
 }
 
 function lastSeenTimeChanged(pub) {
+  console.log('lastSeenTimeChanged', pub, chats[pub].theirLastSeenTime);
   setLatestSeen(pub);
   setDeliveredCheckmarks(pub);
   if (pub !== 'public' && pub === activeChat) {
@@ -636,5 +418,4 @@ function init() {
   $('#scan-chatlink-qr-btn').click(scanChatLinkQr);
 }
 
-export {NewChat, init, showChat, activeChat, chats, addChat, deleteChat, showNewChat, newChat};
-export default ChatView;
+export { init, showChat, activeChat, chats, addChat, deleteChat, showNewChat, newChat, lastSeenTimeChanged, setLatestSeen, setLatestCheckmark};
