+++ src/js/Main.js
@@ -4,8 +4,10 @@ import Helpers from './Helpers.js';
 import PeerManager from './PeerManager.js';
 import Gallery from './Gallery.js';
 import Session from './Session.js';
-import Settings, {LogoutConfirmation, init as initSettings} from './Settings.js';
-import ChatView, {NewChat, chats, init as initChat} from './Chat.js';
+import Settings, {LogoutConfirmation, init as initSettings} from './Components/Settings.js';
+import {chats, init as initChat} from './Chat.js';
+import NewChat from './components/NewChat.js';
+import ChatView from './components/ChatView.js';
 import PublicMessages from './PublicMessages.js';
 import Profile from './Profile.js';
 import QRScanner from './QRScanner.js';
