+++ src/js/components/Settings.js
@@ -1,8 +1,8 @@
-import {publicState, resetView} from './Main.js';
-import Session from './Session.js';
-import Helpers from './Helpers.js';
-import {translate as t} from './Translation.js';
-import { html } from './lib/htm.preact.js';
+import {publicState, resetView} from '../Main.js';
+import Session from '../Session.js';
+import Helpers from '../Helpers.js';
+import {translate as t} from '../Translation.js';
+import { html } from '../lib/htm.preact.js';
 
 const Settings = () => html`
   <div class="main-view" id="settings">
